코드 리팩토링 추가 : 파일 삭제 구현 추가 : 대화상자의 기본 부모 클래스 구현 시작 추가 : 파일 속성 변경 UI 추가

+++ HyperCommander/UIs/Cate - FileOperation/dlgFileProgress.cpp
@@ -79,7 +79,7 @@ void QFileProgress::OnChangedItem( const QString& Src, const QString& Dst )
     ui.lblDestination->setText( Dst );
 }
 
-void QFileProgress::OnChangedProgress( qint16 Item, qint64 Total )
+void QFileProgress::OnChangedProgress( qint16 Item, qint16 Total )
 {
     ui.progItem->setValue( Item );
     ui.progTotal->setValue( Total );
